base functionality for uploadFile
